requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
boto3==1.28.36
botocore==1.31.36
cachetools==4.2.4
channels==4.0.0
decorator==4.4.2
Django==4.2.4
django-bootstrap5==23.3
django-kronos==1.0
django-mathfilters==1.0.0
django-widget-tweaks==1.5.0
jmespath==1.0.1
lru-dict==1.2.0
orjson==3.9.5
platformshconfig==2.4.0
pydantic==1.10.12
python-crontab==3.0.0
python-dateutil==2.8.2
s3transfer==0.6.2
shortuuid==1.0.11
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==1.26.16
